rename page to _page
https://coveord.atlassian.net/browse/KIT-3450
