+++ packages/samples/headless-ssr-commerce/app/_components/_pages/listing-page.tsx
