+++ packages/samples/headless-ssr-commerce/app/_components/_pages/search-page.tsx
